requirements.txt
fastapi
uvicorn
openai
python-dotenv
Jinja2
itsdangerous